requirements.txt
# Deep Learning Framework
# torch
# torchvision>=0.10.0
# torchaudio>=0.9.0

# Transformers và ML Libraries
transformers
scikit-learn

datasets>=2.12.0
tokenizers>=0.13.3
scikit-learn>=1.0.2
numpy>=1.21.0
pandas>=1.4.0
scipy>=1.7.0

# Ollama Dependencies
requests 
websockets>=11.0.3
aiohttp>=3.8.5
asyncio>=3.4.3
litellm>=0.12.0  # For LLM integration

# API và Server
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
python-multipart>=0.0.6
httpx>=0.24.0

# Metrics và Evaluation
seaborn>=0.11.2
matplotlib>=3.5.0
tensorboard>=2.11.0
mlflow>=2.3.0
wandb>=0.15.0

# Model Serving
flask>=2.0.0
gunicorn>=20.1.0

# Data Processing
pillow>=9.0.0
opencv-python>=4.7.0
albumentations>=1.3.0

# Utils
tqdm
pyyaml>=6.0
python-dotenv>=1.0.0
boto3>=1.26.0
google-cloud-storage>=2.9.0
azure-storage-blob>=12.16.0
huggingface_hub>=0.16.4
accelerate>=0.21.0

# Testing và Code Quality
pytest>=7.3.1
pytest-cov>=4.1.0
pytest-asyncio>=0.21.1
black>=23.3.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.4.1

# Optimization
optuna>=3.2.0
ray[tune]>=2.5.0

# Monitoring và Logging
prometheus_client>=0.17.1
grafana-api>=1.0.3
python-logging-handlers>=0.1.0

# Security
cryptography>=41.0.1
python-jose>=3.3.0
passlib>=1.7.4

# Documentation
sphinx>=6.2.1
sphinx-rtd-theme>=1.2.0
mkdocs>=1.4.3
mkdocs-material>=9.1.18

# Development Tools
pre-commit>=3.3.3
jupyterlab>=4.0.2
notebook>=7.0.0
ipython>=8.14.0

# Containerization và Deployment
docker>=6.1.0
kubernetes>=26.1.0

# Optional: GPU support
# Uncomment if using GPU
# cupy-cuda11x>=11.0.0
# torch-cuda>=2.0.0
